risc0_build
Function embed_method_metadata_with_options

pub fn embed_method_metadata_with_options(
    guest_pkg_to_options: HashMap<&str, GuestOptions>,
) -> Vec<MinGuestListEntry>

Build methods for RISC-V and embed minimal metadata - the elf name and path. To embed the full elf, use embed_methods_with_options.

Use this option if you wish to import the guest elf into your prover at runtime rather than embedding it into the prover binary. This reduces build times for large binaries, but makes prover initialization fallible.

Specify custom options for the package by defining its GuestOptions.